+++ nym-vpn-core/crates/nym-gateway-directory/src/entries/gateway.rs
@@ -1,29 +1,15 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::sync::Arc;
-use std::time::{Duration, Instant};
-
-use futures::{SinkExt as _, StreamExt as _};
 use itertools::Itertools;
 use nym_sdk::mixnet::NodeIdentity;
 use nym_topology::IntoGatewayNode;
 use rand::seq::IteratorRandom;
-use rand::{seq::SliceRandom, Rng};
-use tokio::net::TcpStream;
-use tracing::{debug, error, info, trace, warn};
+use tracing::error;
 
-use crate::{error::Result, AuthAddress, Country, Error, IpPacketRouterAddress};
-
-use tokio_tungstenite::connect_async;
-use tokio_tungstenite::{MaybeTlsStream, WebSocketStream};
-use tungstenite::Message;
-type WsConn = WebSocketStream<MaybeTlsStream<TcpStream>>;
-
-const CONCURRENT_GATEWAYS_MEASURED: usize = 20;
-const MEASUREMENTS: usize = 3;
-const CONN_TIMEOUT: Duration = Duration::from_millis(1500);
-const PING_TIMEOUT: Duration = Duration::from_millis(1000);
+use crate::{
+    error::Result, latency_measurement, AuthAddress, Country, Error, IpPacketRouterAddress,
+};
 
 #[derive(Clone, Debug)]
 pub struct Gateway {
@@ -302,153 +288,8 @@ impl GatewayList {
 
     pub(crate) async fn random_low_latency_gateway(&self) -> Result<Gateway> {
         let mut rng = rand::rngs::OsRng;
-
-        let gg: &[Gateway] = &self.gateways;
-        random_low_latency_gateway2(gg).await
-
-        // let gateways_with_latency = Arc::new(tokio::sync::Mutex::new(Vec::new()));
-        // futures::stream::iter(gg)
-        //     .for_each_concurrent(CONCURRENT_GATEWAYS_MEASURED, |gateway| async {
-        //         let id = gateway.identity.clone();
-        //         trace!("measuring latency to {id}...");
-        //         match measure_latency(&gateway).await {
-        //             Ok(with_latency) => {
-        //                 debug!("{id}: {:?}", with_latency.latency);
-        //                 gateways_with_latency.lock().await.push(with_latency);
-        //             }
-        //             Err(err) => {
-        //                 warn!("failed to measure {id}: {err}");
-        //             }
-        //         };
-        //     })
-        //     .await;
-        //
-        // let gateways_with_latency = gateways_with_latency.lock().await;
-        // let chosen = gateways_with_latency
-        //     .choose_weighted(&mut rng, |item| 1. / item.latency.as_secs_f32())
-        //     .expect("invalid selection weight!");
-        //
-        // info!(
-        //     "chose gateway {} with average latency of {:?}",
-        //     chosen.gateway.identity, chosen.latency
-        // );
-        //
-        // Ok(chosen.gateway.clone())
+        latency_measurement::choose_gateway_by_latency(&mut rng, &self.gateways).await
     }
-}
-
-pub(crate) async fn random_low_latency_gateway2(gateways: &[Gateway]) -> Result<Gateway> {
-    let mut rng = rand::rngs::OsRng;
-
-    let gateways_with_latency = Arc::new(tokio::sync::Mutex::new(Vec::new()));
-    futures::stream::iter(gateways)
-        .for_each_concurrent(CONCURRENT_GATEWAYS_MEASURED, |gateway| async {
-            // let id = gateway.identity.clone();
-            // trace!("measuring latency to {id}...");
-            match measure_latency(gateway).await {
-                Ok(with_latency) => {
-                    // debug!("{id}: {:?}", with_latency.latency);
-                    gateways_with_latency.lock().await.push(with_latency);
-                }
-                Err(err) => {
-                    // warn!("failed to measure {id}: {err}");
-                }
-            };
-        })
-        .await;
-
-    let gateways_with_latency = gateways_with_latency.lock().await;
-    let chosen = gateways_with_latency
-        .choose_weighted(&mut rng, |item| 1. / item.latency.as_secs_f32())
-        .expect("invalid selection weight!");
-
-    info!(
-        "chose gateway {} with average latency of {:?}",
-        chosen.gateway.identity, chosen.latency
-    );
-
-    Ok(chosen.gateway.clone())
-}
-
-async fn connect(endpoint: &str) -> Result<WsConn> {
-    match tokio::time::timeout(CONN_TIMEOUT, connect_async(endpoint)).await {
-        // Err(_elapsed) => Err(ClientCoreError::GatewayConnectionTimeout),
-        Err(_elapsed) => todo!(),
-        // Ok(Err(conn_failure)) => Err(conn_failure.into()),
-        Ok(Err(conn_failure)) => todo!(),
-        Ok(Ok((stream, _))) => Ok(stream),
-    }
-}
-
-struct GatewayWithLatency<'a> {
-    gateway: &'a Gateway,
-    latency: Duration,
-}
-
-impl<'a> GatewayWithLatency<'a> {
-    fn new(gateway: &'a Gateway, latency: Duration) -> Self {
-        GatewayWithLatency { gateway, latency }
-    }
-}
-
-async fn measure_latency(gateway: &Gateway) -> Result<GatewayWithLatency> {
-    let addr = gateway.address.unwrap();
-    trace!(
-        "establishing connection to {} ({addr})...",
-        gateway.identity
-    );
-    let mut stream = connect(&addr).await?;
-
-    let mut results = Vec::new();
-    for _ in 0..MEASUREMENTS {
-        let measurement_future = async {
-            let ping_content = vec![1, 2, 3];
-            let start = Instant::now();
-            stream.send(Message::Ping(ping_content.clone())).await?;
-
-            match stream.next().await {
-                Some(Ok(Message::Pong(content))) => {
-                    if content == ping_content {
-                        let elapsed = Instant::now().duration_since(start);
-                        trace!("current ping time: {elapsed:?}");
-                        results.push(elapsed);
-                    } else {
-                        warn!("received a pong message with different content? wtf.")
-                    }
-                }
-                Some(Ok(_)) => warn!("received a message that's not a pong!"),
-                Some(Err(err)) => return Err(err.into()),
-                // None => return Err(ClientCoreError::GatewayConnectionAbruptlyClosed),
-                None => todo!(),
-            }
-
-            Ok::<(), crate::Error>(())
-            // Ok(())
-        };
-
-        let timeout = tokio::time::sleep(PING_TIMEOUT);
-        tokio::pin!(timeout);
-
-        tokio::select! {
-            _ = &mut timeout => {
-                warn!("timed out while trying to perform measurement...")
-            }
-            res = measurement_future => res?,
-        }
-    }
-
-    let count = results.len() as u64;
-    if count == 0 {
-        todo!();
-        // return Err(ClientCoreError::NoGatewayMeasurements {
-        //     identity: gateway.identity_key.to_base58_string(),
-        // });
-    }
-
-    let sum: Duration = results.into_iter().sum();
-    let avg = Duration::from_nanos(sum.as_nanos() as u64 / count);
-
-    Ok(GatewayWithLatency::new(gateway, avg))
 }
 
 impl IntoIterator for GatewayList {
