+++ nym-vpn-core/crates/nym-gateway-directory/src/lib.rs
@@ -5,6 +5,7 @@ mod entries;
 mod error;
 mod gateway_client;
 mod helpers;
+mod latency_measurement;
 
 pub use crate::{
     entries::{
